v1.0.12 && v1.0.13
———————————1.0.13(TEST v1.0.0)  ——————————

- Çıkış işlemine “Çıkış yapmak istediğinize emin misiniz ?” şeklinde uyarı eklendi.
- Android — Bildirim iconu güncellendi.

———————————1.0.12(TEST v1.0.0)  ——————————

- Kullanıcının bildirim sesi seçebilmesi için farklı bildirim sesleri özelliği eklendi.
- Çoklu dil desteği özelliği kapatıldı.
- Çıkış yapma işlemine backend servisi entegre edildi.

+++ lib/features/login/presentation/cubit/auth_cubit/auth_cubit.dart
@@ -50,24 +50,24 @@ final class AuthCubit extends Cubit<AuthState> with BaseCubit<AuthState> {
   }
 
   Future<void> logout() async {
-    _loginLocalService.delete(CacheConstants.User.name);
-    AppRouter.goNamed(AppRoutes.Login.path);
+    // _loginLocalService.delete(CacheConstants.User.name);
+    // AppRouter.goNamed(AppRoutes.Login.path);
 
-    // safeEmit(AuthState.loading());
+    safeEmit(AuthState.loading());
 
-    // final response = await _loginRepository.logout();
+    final response = await _loginRepository.logout();
 
-    // final value = switch (response) {
-    //   Success(value: final bool isLogout) => {
-    //       if (isLogout)
-    //         {
-    //           _loginLocalService.delete(CacheConstants.User.name),
-    //           AppRouter.goNamed(AppRoutes.Login.path),
-    //         }
-    //     },
-    //   Failure(exception: final AppException exception) => {
-    //       showErrorBottomSheet(exceptionMessage: exception.message),
-    //     }
-    // };
+    final value = switch (response) {
+      Success(value: final bool isLogout) => {
+          if (isLogout)
+            {
+              _loginLocalService.delete(CacheConstants.User.name),
+              AppRouter.goNamed(AppRoutes.Login.path),
+            }
+        },
+      Failure(exception: final AppException exception) => {
+          showErrorBottomSheet(exceptionMessage: exception.message),
+        }
+    };
   }
 }
